+++ src/views/chat/components/Message/Text.vue
@@ -1,9 +1,10 @@
 <script lang="ts" setup>
-import { computed } from 'vue'
+import { computed, ref } from 'vue'
 import { marked } from 'marked'
 import hljs from 'highlight.js'
 import { useBasicLayout } from '@/hooks/useBasicLayout'
 import { encodeHTML } from '@/utils/format'
+import { t } from '@/locales'
 
 interface Props {
   inversion?: boolean
@@ -18,14 +19,18 @@ const { isMobile } = useBasicLayout()
 
 const renderer = new marked.Renderer()
 
+const textRef = ref<HTMLElement>()
+
 renderer.html = (html) => {
   return `<p>${encodeHTML(html)}</p>`
 }
 
 renderer.code = (code, language) => {
   const validLang = !!(language && hljs.getLanguage(language))
-  if (validLang)
-    return `<pre><code class="hljs ${language}">${hljs.highlight(language, code).value}</code></pre>`
+  if (validLang) {
+    const lang = language ?? ''
+    return `<pre class="code-block-wrapper"><div class="code-block-header"><span class="code-block-header__lang">${lang}</span><span class="code-block-header__copy">${t('chat.copyCode')}</span></div><code class="hljs code-block-body ${language}">${hljs.highlight(lang, code).value}</code></pre>`
+  }
   return `<pre style="background: none">${hljs.highlightAuto(code).value}</pre>`
 }
 
@@ -54,6 +59,8 @@ const text = computed(() => {
     return marked(value)
   return value
 })
+
+defineExpose({ textRef })
 </script>
 
 <template>
@@ -62,7 +69,7 @@ const text = computed(() => {
       <span class="dark:text-white w-[4px] h-[20px] block animate-blink" />
     </template>
     <template v-else>
-      <div class="leading-relaxed break-all">
+      <div ref="textRef" class="leading-relaxed break-all">
         <div v-if="!inversion" class="markdown-body" v-html="text" />
         <div v-else class="whitespace-pre-wrap" v-text="text" />
       </div>